requirements.txt
requests
python-dotenv
litellm
logging
